lb1/pyproject.toml
[project]
name = "lb1"
version = "0.1.0"
requires-python = ">=3.14"
dependencies = [
    "django>=6.0.3",
    "djangorestframework>=3.17.1",
]
